refactor(ReservationListFragment): MVP 리팩터링
