+++ app/src/main/java/woowacourse/movie/view/moviemain/reservationlist/ReservationListContract.kt
@@ -1,0 +1,9 @@
+package woowacourse.movie.view.moviemain.reservationlist
+
+import woowacourse.movie.view.model.ReservationUiModel
+
+interface ReservationListContract {
+    interface Presenter {
+        fun getReservations(): List<ReservationUiModel>
+    }
+}
