+++ app/src/main/java/woowacourse/movie/view/moviemain/reservationlist/ReservationListFragment.kt
@@ -5,15 +5,13 @@ import android.view.View
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
-import woowacourse.movie.data.ReservationMockRepository
 import woowacourse.movie.view.ReservationCompletedActivity
-import woowacourse.movie.view.mapper.toUiModel
 
 class ReservationListFragment : Fragment(R.layout.fragment_reservation_list) {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         val recyclerView = view.findViewById<RecyclerView>(R.id.recyclerview)
-        val reservations = ReservationMockRepository.findAll().map { it.toUiModel() }
+        val reservations = ReservationListPresenter().getReservations()
         recyclerView.adapter = ReservationListAdapter(reservations) { reservation ->
             val intent = ReservationCompletedActivity.newIntent(requireContext(), reservation)
             startActivity(intent)
